test: do not immediately fail tests on 404
The app is returning `404` errors when using routes to navigate directly
to a view e.g. `<base-url>/profile`, this is an infrastructure
limitation and the response body is still the full working application.
Update the Cypress tests to no longer fail based on the status code when
navigating directly to a specific view. Use the base url wherever
appropriate to reduce the number of places we need to ignore the status
code.

TIS21-3726