chore: invalidate cloudfront cache on deploy
The static JS and CSS content is cached within Cloudfront, when
deploying a new version of the site that cache should be invalidated.
Technically this is not required for normal running of the application,
as the CSS an JS files are versioned with a content hash. While
Cloudfront does not need to re-cache a file with the same name,
invalidating the cache does mean that old un-used version are cleaned
up more quickly.

TIS21-3765
TIS21-3781