test: More Jets config and remove Enzyme
Jest throws 'unexpected token...' error which was fixed by adding more
config to the test script - see https://stackoverflow.com/questions/62328584/react-jest-test-fails-to-run-with-ts-jest-encountered-an-unexpected-token-on-i
for more info.

Enzyme is now dead so removed those Jest tests that use it. Some will
need redoing as either Cypress Component Tests or React Testing Library
tests - seehttps://dev.to/wojtekmaj/enzyme-is-dead-now-what-ekl for
more info.