chore: add cypress e2e instrumentation
trying to get instrumentation and code coverage for the e2e tests in cypress,

we are using '@istanbuljs/nyc-config-typescript'  with 'babel-plugin-istanbul'

for the nextjs config we are using these examples:

https://github.com/kylemh/next-ts-with-cypress-coverage

https://github.com/bahmutov/next-and-cypress-example

we have the .nyc_output folder which appears to have the instrumented code within. the cypress tests are not displaying a report.