fix: form R B confirm submission bug
The bug is on the declarations page and occurs when the declarations have been ticked and a draft form is saved. Instead of just saving the form with the accompanying toast message the confirm modal for form submission also appears. This is due to the formik onsubmit logic running incorrectly when the the declarations are ticked. The fix involves resetting the form values on 'save draft' click event to ensure the save draft logic runs and a draft is saved without a confirm modal.