fix: stale user state
The switch to Next js has uncovered an existing Amplify issue where changes to the Amplify user provided by the Authenticator component does not a trigger an update to the user object (e.g. see https://github.com/aws-amplify/amplify-ui/issues/1333 ).
This results in 'stale' user data e.g. user.preferredMFA not being updated from NOMFA when MFA type is updated.
The solution is to use Auth.currentAuthenticatedUser() and the Redux store to ensure the latest user data is made available to all the components. This also simplifies things a bit by not having to pass the user and mfa props down several components.