fix: form R B confirm submission bug
The bug is on the declarations page and occurs when the declarations have been ticked and a draft form is saved. Instead of just saving the form with the accompanying toast message, the confirm modal for form submission also appears.
The quick fix involves resetting the form values on 'save draft' click event to prevent form submission but allow a saved draft (minus a confirm modal).