feat: Add basic DSP comps and pre-issue logic
Store current path in local storage via custom useLocalStorage hook - to be used to redirect user after DSP cred has been added to wallet.
Fix Placement and ProgrammeMembership models to allow signatures to be captured and stored.
Create function to store Profile data object matched with tisId of clicked panel (to be used in cred issuing)
Refactor panel generator to allow tisId to be available for the above. Generator panel could so with further work (e.g. current form means that type safety is lost at the individual panel level. The only way I can see of fixing this at the moment is to separate out the logic for panels and placements.)
Made a start on the CredentialsService but will need the FS'ers help.