feat: move `COJ_EPOCH` value to env configs
The `COJ_EPOCH` is the date that programme membership start dates are
compared to when deciding whether to allow signing of the CoJ via the
app. Currently, this is targeted for 2023-08-01 to enable the August
rotations to use TSS for CoJ.
However, this would leave local and preprod environments unable to test
the functionality until we get within 13 weeks of August.

Update the environment files to include a variable
`NEXT_PUBLIC_CONDITIONS_OF_JOINING_EPOCH` with some sensible defaults
for each environment.
Update the `COJ_EPOCH` constant to build a date from the new environment
variable.

TIS21-4384
TIS21-4373