test: CojUtilities.getStatusText tests
The refactoring of the `CojUtilities.getStatusText()` tests to handle
the addition of the "available from" message has left those tests in a
broken state.
One test passes now but will fail starting 2023-08-01, the other is more
stable but the description and expectation no longer match.

Refactor/rewrite the tests for date >= COJ_EPOCH scenarios.

TIS21-4386
TIS21-4374