refactor: `ConditionsOfJoining` component
Rename the component from `ConditionsOfJoiningField` to
`ConditionsOfJoining` import the model as `ConditionsOfJoiningModel` to
facilitate this.
Update the `PanelsCreator` to use the new component name.

Refactor the status checks to use `CojUtilities.canBeSigned()` instead
of checking against the text value of `getStatusText()`. Remove the
dummy link for viewing the form, that will be added by a future ticket.
Change the button to a summary list action, as per the NHS design
manual.

Update `CojUtilities` to make `canBeSigned` public, it should also check
against the COJ_EPOCH to avoid allowing signing of programmes starting
too early.

TIS21-4387
TIS21-4374