refactor: style COJ actions as links
The COJ actions are currently a button, but should be styled as links
instead.
Swap the `button` element for `Link`s and refactor the `viewCoj`
function in to `setCojState` with the navigation done within the link
itself.

Update the ConditionOfJoining component tests to wrap the component in a
`BrowserRouter` so that the `Link`s can work during testing.

TIS21-4544
TIS21-4533