fix: JSON form fields bug
Fix the form fields logic so field changes are reset when a new form is opened.(e.g. show/hide changes do not directly mutate the form object and so do not persist when navigating away and then opening a new form).
Make a start on refactoring the handleChange function to make it more readable (cognitive complexity code smell).