fix: the form state
The last merge to main did reset the json form fields to their defaults when navigating away from the form but had numerous unintended consequences. This commit fixes those issues and bring the formBuilder back to the state before the "fix".
The Cypress tests have also been updated to:
1. remove the otherImmigrationStatus test from FormRA as the "other" immigration options are not on prod.
2. Update the formsList btn test.

TODO
1. Fix the bug where fields made visible by a conditional are not reset when navigating away from the form.