test: Add test for the bug fix
Added a test to check that the app correctly persists the updated dependent field visibility status to trigger any expected validation when a draft form is re-opened.

Tried to and failed to get Cypress set-up working to test the condition where the the status of a changed dependent field's visibility is not persisted when navigating away from the form and opening a new one. One for a follow-up ticket I think.