fix: Form B bug
...when navigating to next page mid-autosave.
If you are quick then you can make a change on Form B and then nav to next section before the autosave completes (it won't complete until it is triggered again).
It's only an issue with Fomr B so is likely to be temporary once we refactor to use form builder.
So for now I've disabled the 'next' nav Link (on both forms for consistency) when autosaving is in progress.
I've also reduced time for autosave to kick in from 5s to 2s again to reduce the chance of this bug happening.
Will have to test the UX with some of our volunteers to see if this is not too annoying.