feat: swap position of buttons and pagination
Makes more sense to have the pagination links above the 'save' and 'start over' buttons as they are used more often.
Also added 'GMT' (felt dirty doing it 😅) to the submitted form time to lessen any confusion between this time and the autosave timestamp time. (Would be good if we could save timezone info to Mongo. 🥫🐍<= best I can do for a worm.)