refactor: bookmark alert
Refactor the bookmark alert to be based on the redux store, when the
page is loaded the URL will be checked for `redirected=1` and the stored
value set appropriately.

The `redirected` parameter should be cleaned from the URL to avoid it
being included in the bookmark and triggering the alert every visit.

Update the bookmark alert to be more explicit about what has changed.

TIS21-4887
TIS21-4784