fix: Global Alert start over issue
Update form status back to idle to trigger a re-render to refresh the form list state so the correct message is in the alert banner when the draft form is deleted.
Again, not the prettiest fix but it works.
Need to have a proper look at form-related state management across the app and when to trigger data-fetch and component re-renders.