fix: Global Bookmark Alert visibility (NOMFA)
In line with current behaviour, make the Bookmark alert visible even when a user has yet to set their MFA.
The Action Summary will be visible when the user has set their MFA and is redirected to the home page. The Global Action Summary alert will then be visible when user navigates to other pages (assuming they have actions to view).

Also added tests for this in GlobalALert.cy.tsx and tidied up this test file to make it easier to follow.