feat: re-enable covid declaration flag
this has been re-enabled because when a form has been unsubmitted or is in draft and the user has already compleated the covid declaration section before the switch, the user will lose the ability to alter their covid section and will be stuck with their previous changes. which they will also still be able to view.