WIP #10 Continue with form errors logic
Main focus for this commit is to display the inline errors for array panel fields when the validateCurrentField function is called (on Text, Selector, and Dates for now).
Had to alter the vaidation schema a bit to get access to the parent object to compare the work.startDate and endDate when endDate onChange.
TODO: With the validateCurrentField logic only validating the current field, this does leave open the possiblity of 'stale' error messages for linked fields (e.g. startDate and endDate). This will be addressed in a future commit. There are 2 options I can think of atm, either change the FormBuilder logic to validate the whole page when a field changes (as is the case with the existing Form B Formik implementation) or add a 'validateLinkedFields' function that can be called when a field that has a 'validatelinkedFields' property (via the JSON form fields file?) is updated by the user.
Note: Also added a native date input field component (DatesNative.tsx) as part of the work to fix the work.endDate validation which I didn't end up using. But have kept it for now as it may be useful if we decide to ditch the react-datepicker library (see https://hee-tis.atlassian.net/browse/TIS21-5653 for more info).