WIP #11 - rework validation
Following discussion around validating current field limitations (linked fields e.g. startDate and endDate), it was decided to validate the whole page via the handleChange rather than just the current field. This is the current behaviour for Form B.

So far, reworked the handleChange to validate the whole page and display the errors for non-array fields.