WIP #14 Work on confirm logic
Move handle whole-form visibility logic to the view builder
Re-work formbuilder fields and formData logic as a result.

Still need to accomodate panels in the view builder, errors summary and a few other bits and pieces (see FormBuilder TODO comments).

Rename formFields to formData so it is not confused with fields (JSON schema fields) var.