feat: Add Declaration component
...to handle the submission declarations.
This bit of the form is a bit tricky to standardise as the declarations for Form A are not coupled to the form data while those for Form B are.
The primary purpose for this commit is to create a component which can take in the as props the checkbox declarations to display before submission.
The logic to add any declarations to the form data will be handled in the form submission function.