feat: Add Declarations checklist component
There is currently no consistency with Form R submission declarations.
Standardise the submission checklist so the checklist fields and text can be passed in via the JSON file
The logic for checkbox and ‘Submit Form’ state (e.g. disabled if not all checkboxes have been checked) is also self-contained in a Declarations.tsx component.

Note: A large portion of the submit logic added to FormAView.tsx has been copied from the existing logic in formr-part-b/sections/Declarations.tsx .
There's still scope to refactor this code too but the main priority is to have something that can be used to build both forms.