test: Notifications components (Cypress comp)
Test the basic table interactions (sort, filter, pagination).
No tests yet for the 'mark as unread' and row click actions.
Test the messageText and one error condition for the message component.
No tests yet for the 'success' conditon for the message component.

Note: Cypress is behaving strangely with the 'mark as unread'/row click actions - not sure if it's Cypress or my code.
I know which I'd put money on though 😅
Will need to investigate further. Might need to refactor the code to make it more testable (In Particular, focus on the RowActions.tsx and NotificationsTable row click events that call the updateNotificationStatus function).