test: 'mark as unread' btn and row click
Use RTL (Jest) so we can mock the updateNotificationStatus utils function.
Move more logic from the components to the utils file and split the utils functions into smaller ones - which may make things more testable!

Still need to test the full flow of the mark as unread button and row click - if anyone fancies it!