fix: improve in-app notification formatting
The in-app notification message contents is rendered as plain text,
making them difficult to read.

The notification message will be modified to return as a HTML fragment,
so the notification message component should be modified to use the
fragment as inner html.

Add DOMPurify as a dependency and use it to sanitize the fragment before
inserting it in to the DOM.

TIS21-5898
TIS21-5900