feat: Add shortcut to confirm btn to builder...
...to mirror the current FormB behaviour when editing.
chore: Add about half the logic to enable the FormBuilder to build FormB and not break much.

NOTE: This is slice 1 of at least 2 commits to build FormB via the FormBuilder. Hopefully, this breaks things up into more reviewable chunks.
There is a fully-working FormB (via the FormBuilder) but it's a beast of a commit: https://github.com/Health-Education-England/tis-trainee-ui/tree/feat/BobTheFormBuilder_partB.

The main focus on this commit is to add the necessary logic to the FormBuilder so it can build any complex form (e.g. FormB), and for FormA to still be fully-working and tested.
The idea would be to push this to preprod, then add the next commit(s) there to get Form B working via the builder.
The trickiest logic was around the COVID declaration dto field. Most straightforward solution to solve the immediate problem is to create a new "dto" FieldType (see FormBuilder.tsx and README for other types) and DtoBuilder. Longer term, we could look at a more recursive solution.
More details about Form B logic to follow.

Changes include:
1. Consolidate the form component logic into a single FormRPartA.tsx file (routes plus data). A new FormrPartB.tsx will follow.
2. A new DtoBuilder.tsx to handle the new dto field type.
3. A new FormView.tsx to handle any form View/Confirm page.
4. A new ExpanderMsg.tsx to handle any expander text (not just dataSource).
5. Some new Cypress commands to make life a bit less repetitive.

Most of the other changes are just refactoring to make the above work for any form built with FormBuilder.tsx.
There is also a bit of refactoring as far as the field visibility logic goes in order to handle all the field types.
e.g. see FormBuilder.tsx. where the useEffect that tracks the field visibility has been deleted and the logic moves to the field rendering.
e.g. see FormBuilder.tsx, PanelBuilder.tsx, and DtoBuilder.tsx for showFormField(), and FormViewBuilder.tsx: for VisibleField JSX to recursively handle dto fields rendered in the View.