fix: unable to record one day locum
A single day locum shift will start and end on the same day, currently
that fails the Form R validation rules.

Update the validation to allow the start date and end date to be the
same, it should still fail if the end date is before the start date.

TIS21-6084