fix: unsubmitted form declaration type
Currently, for a new DRAFT Form R B, the currentDeclarations and previousDeclarations initial values are a blank array.
However, when the user submits the form any blank array is saved as null. This results in a bug when the user is editing an UNSUBMITTED form where the app will crash because it is still expecting a blank array.

Fix the bug by adding some additional logic to handle the null type for currentDeclarations and previousDeclarations.
Also inititialise these 2 fields as null for new DRAFT forms (as it's easier to test the new logic!).