fix: default dialog position
Use window.scrollY value to align the dialog with the y scroll position so the CCT calc is in view when/if a btn click happens down the page.
This involves using defaultPosition prop rather than position - which then means we can't use position prop to reset the dialog to 0,0 when printing (clicking 'save PDF').
i.e. defaultPosition only works on first render.
Best solution is to make a separate PrintableCct comp using basic HTML that we can position and render more consistently. This involves only rendering it when shouldPrint is true and also keeping the contents hidden except when printing (via @media print) etc.
test: Update test to reflect changes - although not really possible to test some of the logic around the PrintableCct rendering etc. via the 'save PDF' click/ print preview.