refactor: action summary alert
Update logic so it appears on all but action summary page.
Move the global alert to the top of the page - and reduce its size slightly - to help improve visibility of the main pages.
Simplify the global alert comps html and add icons to action summary alert msgs.

TODO: fix the tests