feat: Add 'New Prog' onboarding journey diagram
...for mobile and desktop views.
Initially developed the 'tracker' assuming active statuses as per ACs so there is quite a bit of
redundant action status logic lying around for when the API is ready.
(Could refactor to remove in a future commit if this is preferred.)
Also, allow user to click on info icon in a given action card, to open a modal displaying links/further details.
Allow user to close modal via 'close' btn, escape key, or by clicking outside modal.
(Note: The CCT calc does not currently close via click outside as it was developed before the Modal comp so needs refactoring.)