fix: eslint error in Analyse PR
This error is in Analyse PR action but not locally (prob due to diff in versions. Might be time for an update!):

[{"ruleId":"react-hooks/exhaustive-deps","severity":1,"message":
"The ref value 'modalRef.current' will likely have changed by the time this effect cleanup function runs.
If this ref points to a node rendered by React, copy 'modalRef.current' to a variable inside the effect,
and use that variable in the cleanup function.","line":67,"column":16,"nodeType":"Identifier","endLine":67,"endColumn":23}]