feat: Add CCT calc (for single LTFT change)
As agreed, FE handles the cctDate calc /create -> /view.
The BE will still calculate cctDate for form POST and PUT - and also GET for returning a saved calc (all via /cct/calculation/{id}).
CctCalcView is still WIP (TODO more error-handling, redirects) as are some more tests.

ps. Apologies for size of commit but with all the toing and froing it was easier to refactor and force push as one gelatinous blob.