fix(form-r): print to PDF
The Form-Rs print to PDF functionality is outputting blank pages due to
a CSS change which inadvertently hides the body.
There is a `pdf-visible` class which is intended as a flag for what
should be printed, but it has not yet been added to the Form-R views.

Remove the CSS change until the Form-R views can be properly flagged as
PDF visible.

TIS21-6799