chore: move formData to context
This is the first step to separate the FormBuilder state management from the rendering.
The idea is to use a Provider to manage the form state and have the FormBuilder components access the state via the context.
This will also make managing the LTFT form (and any future forms) state easier.