fix: CctCalcSummaryDetails
The original cctSnapshot works well when the snapshot is in the redux store.
However, the snapshot is not persisted and is empty after e.g. page refresh, saved reload, etc.
Instead, get the cct calc data directly from the formData.

TODO: Some FE/BE DTO disparities when returning a saved form:
change: cctDate is null
personalDetails: id field should not be here, skilledWorkerVisaHolder should still be Type null if radio not yet selected but is returned false.