fix: startover btn visibility on form view page
Fix bug where the startover button is not visible on the form view page
if e.g. the autosave fails as the user completes the form (or they are extremely quick and navigate without the autosave kicking in) up to the point of submission.
The visibility of the startover button on the form view/ submit page has relied on the form having an id
either from the autosave (stored in redux as newFormId) or from the id of a retreived form.
The fix is to add a check for the btnLocation (via a prop in StartOverButton) to ensure the
button is always visible on the form view page. If there is no form id, the button click event then defaults to a simple redirect to the forms list page.