test: LtftFormView (and modal render)
refactor: LtftFormView.cy.tsx to remove the modal render tests as they no longer work with the new save call.
The LtftFormView -> saveDraftForm -> LtftNameModal render logic tests are in LtftFormView.test.tsx.
refactor: LtftNameModal.test.tsx to account for removal of naming logic and the addition of 'submitting' text on button.

test: SaveDraftForm, handleSaveRedirect, and chooseRedirectPath utilities.

TODO: test for the effect of the showFailToastOnly in the ltftSlice Toast msg rendering for saveLtft and updateLtft redux thunks.