fix: Action Modal button isSubmitting state
Pass in isSubmitting state as prop to the ActionModal component to disable the dialog modal btn and display the correct btn text when the form is submitting.
Change the ordering of the modal close so that the submission logic runs first before closing the modal.
This ensures that the dialog modal button is still visible until the submission logic concludes.
It also means that we can delete similar btn disable logic in the RowCctActions as it is now handled in the ActionModal component.

Note: the previous btn disable logic in the dialog modal did not work because the useSubmitting hook in the ActionModal component was not accessing the correct state value set in the parent component and so would always be false.