fix: handlePageChange bug
In the FormBuilder, add isShortcut flag to handlePageChange function to distinguish between ususal navigation
and shortcut back to Confirm page.
fix: shortcut btn rendering so it isn't visible when saved form first loads i.e. the user has not come from the Confirm page to edit the form.
(although this is the as-is behaviour, we may want to change it so the shortcut button is always visible for saved draft?)