fix: form edit state bugs
fix: The form edit state is not being reset correctly when reopening a draft form after editing it.
fix: The ususal next-page navigation is incorrectly shortcutting to the Confirm page after editing.

TODO: Next commit will move the handlePageChange function out of FormBuilder so will write tests for it then.