test: SaveAndExitButton
Create a SaveAndExitButton.tsx component to separate concerns (i.e.
the button's behavior is independent of the form's context) and
make it easier to test all the button conditions.
Add a couple of Cypress Component/e2e tests too to ensure it's working as expected as
the form is being filled out.