fix: FormView test
Most of the test file set-up was fine, main issue is the circular dependency issue in FormBuilderUtilities:
chooseRedirectPath is referenced before initialization because of the way the mock is set up.
(The fix follows a similar set-up pattern to other RTL (Jest) test files - which involves a lot of 'jest.mock'! For examples search for '.test.tsx')

Also imported 'act' directly from "react" because the other one is deprecated.