test: clarify test when no features enabled
The tests for navlink behaviour when no features were enabled are
confusing phrased as being related to unauthenticated users.
That's not a valid app state, so the tests were misleading.

Rename the mock data to `mockUserFeaturesNone` and rephrase the test
descriptions to be specifically when all features are disabled, rather
than being focused on an impossible user state.

TIS21-7657
TIS21-7659
TIS21-7661