test: fix and verify critical success
The `Main` component tests always result in a critical error due to
feature flag and actions endpoints being called and resulting in a 404.
Update `Main.cy.tsx` and the relevant redux slices to allow the feature
and actions statuses to be set. Add an additional test to verify that
the app loads without error when all required data is available.

TIS21-7657