test: cache signed in session
Use Cypress sessions to cache the sign-in state, allowing tests within a
spec to only need to sign in the first time.
This eliminates the 30s wait time between each tests when multiple tests
within a single spec call the sign-in command.

The session cannot be maintained between specs, so we still need the 30
second wait period to account for each spec loggin in separately.

TIS21-5489
TIS21-7953