fix: ‘Anchor used as a button...’ build error
... by using react-router-dom Link instead of anchor.
refactor: split handleEditSection util function logic into separate functions for Link ‘to’ and ‘onClick’ props.

refactor: FormViewBuilder to make it more readable/maintainable.
fix: FormViewBuilder logic to make sure the 'change' Link is still rendered for an empty otherDiscussions array.
This logic is only need for this field due to the way the LTFT form page is constructed.
fix: ArrayFieldRenderer valueType passed into the formatEntryValue function to ensure it is correct for a nested
field within an array. Previously, the formatEntryValue was using the parent array field type which meant
the (fieldType === "date") condition was useless as the fieldType was always "array".
(Probably still catch it with the regex condition but not as robust.)