refactor: move the form initialization logic
... from the FormsListBtn component to the FormRPartAForm component.
This includes the FormLinkerModal rendering.
This allows us to simplify the FormsListBtn component so it just
handles the btn text and navigation.
We can now encapsulate all form-creation/editing logic within the form comp.
This includes direct navigation to the /new/create route.

I think the plan for this work is to get all the formr and ltft forms
logic kinda working first then separate this into a few PRs.