refactor: move the form initialization logic
... from the FormRListBtn component (renamed from FormsListBtn and moved to form-r folder)
to the FormRPartAForm/FormRPartBForm component, which includes the FormLinkerModal rendering.
This allows us to simplify the FormRListBtn component so it just handles the btn text and navigation.
We can now encapsulate all form-creation/editing logic within the form comp.
e.g. direct navigation to e.g. /new/create route.

Note: There is a fair bit of duplication between FormRPartAForm and FormRPartBForm.
Might address in this PR.