refactor: FormR Views
... to make them more independent of the other form pages and allow
direct navigation.
The useEffect conditional logic is trickier than I expected
(e.g. see FormRPartAView.tsx). It not only checks for the id param,
but also if the user is coming from the form creation page (!fromCreate),
and if the form data is available from the store (!formData?.traineeTisId).

The last formData condition can be omitted (together with the appropriate
 refactor of the second return) but this leads to a quick rendering of the
 "Failed to load your Form R..." error before the data page is rendered.
Note: the formData is left out of the useEffect dependencies
to prevent infinite loops (so we use a eslint-disable...).

fix: canEdit lifecycle logic and removed redirect return as the returns are
handled in the parent.