fix: form errors summary display
... and scroll to the particular field that has the error.
This is in line with NHS design system recommendations for forms with errors.

Note: This change is for the main form and view pages.
I have left the form error summary at the bottom of the main form page as it is probably
more useful to have it here as you attempt to navigate the form (but could be revisited later).
TODO: Use the NHS class nhsuk-form-group--error to wrap the the field errors to match the design system.