fix: bug where first View refresh renders error
... if the user has navigated from the main form
and then refreshes the View page (only happens on first refresh).

The issue is caused by a combination of two factors:
1. When refreshing the page, the Redux store is reset,
so formData reverts to its initial state (where lifecycleState is New)
and formLoadStatus is idle.
2. If fromCreate is true (which can persist in location.state across a page refresh),
the data fetch useEffect is skipped as the logic assumes the data is already
in the store via the navigation from the main form (i.e. saving the formData and
setting location state fromCreate === true).

Additionally, even if the fetch is triggered, th error page will initally
render if the status is "idle" and we are about to fetch the data
but the initial form state has a NEW form.

The fix:
1. Update the data fetch useEffect to not skip fetching
if fromCreate is true but initial form state is NEW.
2. Update the error page rendering logic to not render if
we have the Id but the status is "idle" (i.e. we are about to fetch data).